refactor authentication and wrapper methods
