+++ Backend/scala-backend/app/controllers/SchoolClassController.scala
@@ -32,7 +32,6 @@ class SchoolClassController @Inject() (
   implicit val schoolClassWrites = Json.writes[SchoolClassCC]
 
   implicit val classTeacherReads = Json.reads[ClassTeacherCC]
-  implicit val classTeacherWrites = Json.writes[ClassTeacherCC]
 
   private val model = new SchoolClassModel(db)
 
@@ -97,27 +96,4 @@ class SchoolClassController @Inject() (
       })
       
   }
-  // GET /teacherAuth/:id/:classSecret
-  // returns ClassTeacherT, eine abgespeckte version der schoolclass
-  def authenticateTeacher(id: Int, classSecret: String): play.api.mvc.Action[play.api.mvc.AnyContent] = Action.async { implicit request: Request[AnyContent] =>
-    val accepted: Future[Boolean] = model.validateAccess(id, classSecret)
-    accepted.flatMap(a => {
-        if(a) {
-          request.headers.get("teacherSecret") match {
-            case Some(teacherSecret) => {
-                model.getTeacher(id, teacherSecret).map(result => result match {
-                  case Some(teacher) => Ok(Json.toJson(teacher)) //return
-                  case None => Forbidden("Wrong teacherSecret") //return
-                })
-            }
-            case None => Future.successful(BadRequest("No teacherSecret provided")) //return
-          }
-        }else{
-          Future.successful(NotFound("Schoolclass with that id not found or wrong classSecret")) //return
-        }
-     })
-    
-
-  }
-
 }
